+++ src/main/java/com/springboot/monew/common/exception/ErrorResponse.java
@@ -1,4 +1,4 @@
-package com.springboot.monew.exception;
+package com.springboot.monew.common.exception;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
