+++ src/main/java/com/springboot/monew/exception/article/ArticleException.java
@@ -1,7 +1,7 @@
 package com.springboot.monew.exception.article;
 
-import com.springboot.monew.exception.ErrorCode;
-import com.springboot.monew.exception.MonewException;
+import com.springboot.monew.common.exception.ErrorCode;
+import com.springboot.monew.common.exception.MonewException;
 
 import java.util.Map;
